Modified Epigraph Index of univariate functional dataset

This function computes the Modified Epigraphic Index (MEI) of elements of a univariate functional dataset.

MEI(Data)

# S3 method for fData
MEI(Data)

# S3 method for default
MEI(Data)

Arguments

Data

either an fData object or a matrix-like dataset of functional data (e.g. fData$values), with observations as rows and measurements over grid points as columns.

Value

The function returns a vector containing the values of MEI for each element of the functional dataset provided in Data.

Details

Given a univariate functional dataset, \(X_1(t), X_2(t), \ldots, X_N(t)\), defined over a compact interval \(I=[a,b]\), this function computes the MEI, i.e.:

$$MEI( X(t) ) = \frac{1}{N} \sum_{i=1}^N \tilde{\lambda}( X(t) \leq X_i(t) ), $$

where \(\tilde{\lambda}(\cdot)\) is the normalized Lebesgue measure over \(I=[a,b]\), that is \(\tilde{\lambda(A)} = \lambda( A ) / ( b - a )\).

References

Lopez-Pintado, S. and Romo, J. (2012). A half-region depth for functional data, Computational Statistics and Data Analysis, 55, 1679-1695.

Arribas-Gil, A., and Romo, J. (2014). Shape outlier detection and visualization for functional data: the outliergram, Biostatistics, 15(4), 603-619.

See also

Examples

N = 20
P = 1e2

grid = seq( 0, 1, length.out = P )

C = exp_cov_function( grid, alpha = 0.2, beta = 0.3 )

Data = generate_gauss_fdata( N,
                             centerline = sin( 2 * pi * grid ),
                             C )

fD = fData( grid, Data )

MEI( fD )
#>  [1] 0.4760 0.2900 0.2480 0.1545 0.8980 0.4775 0.2970 0.3885 0.7515 0.7395
#> [11] 0.5365 0.7580 0.9800 0.4790 0.0940 0.8430 0.9380 0.0565 0.6515 0.4430

MEI( Data )
#>  [1] 0.4760 0.2900 0.2480 0.1545 0.8980 0.4775 0.2970 0.3885 0.7515 0.7395
#> [11] 0.5365 0.7580 0.9800 0.4790 0.0940 0.8430 0.9380 0.0565 0.6515 0.4430