Modified Hypograph Index of univariate functional dataset

This function computes the Modified Hypograph Index (MEI) of elements of a univariate functional dataset.

MHI(Data)

# S3 method for fData
MHI(Data)

# S3 method for default
MHI(Data)

Arguments

Data
either an fData object or a matrix-like dataset of functional data (e.g. fData$values), with observations as rows and measurements over grid points as columns.

Value

The function returns a vector containing the values of MHI for each element of the functional dataset provided in Data.

Details

Given a univariate functional dataset, \(X_1(t), X_2(t), \ldots, X_N(t)\), defined over a compact interval \(I=[a,b]\), this function computes the MHI, i.e.:

$$MHI( X(t) ) = \frac{1}{N} \sum_{i=1}^N \tilde{\lambda}( X(t) \geq X_i(t) ), $$

where \(\tilde{\lambda}(\cdot)\) is the normalized Lebesgue measure over \(I=[a,b]\), that is \(\tilde{\lambda(A)} = \lambda( A ) / ( b - a )\).

References

Lopez-Pintado, S. and Romo, J. (2012). A half-region depth for functional data, Computational Statistics and Data Analysis, 55, 1679-1695.

Arribas-Gil, A., and Romo, J. (2014). Shape outlier detection and visualization for functional data: the outliergram, Biostatistics, 15(4), 603-619.

See also

Examples

N = 20
P = 1e2

grid = seq( 0, 1, length.out = P )

C = exp_cov_function( grid, alpha = 0.2, beta = 0.3 )

Data = generate_gauss_fdata( N,
                             centerline = sin( 2 * pi * grid ),
                             C )
fD = fData( grid, Data )

MHI( fD )

#>  [1] 0.6890 0.0620 0.1610 0.7260 0.2370 0.7845 0.3770 0.1125 0.4735 0.4870
#> [11] 0.4380 0.6315 0.9070 0.4915 0.3430 0.9750 0.1870 0.6950 0.9530 0.7695

MHI( Data )

#>  [1] 0.6890 0.0620 0.1610 0.7260 0.2370 0.7845 0.3770 0.1125 0.4735 0.4870
#> [11] 0.4380 0.6315 0.9070 0.4915 0.3430 0.9750 0.1870 0.6950 0.9530 0.7695