Maxima of a univariate functional dataset

This function computes the maximum value of each element of a univariate functional dataset, optionally returning also the value of the grid where they are fulfilled.

maxima(fData, ..., which = FALSE)

Arguments

fData

the functional dataset containing elements whose maxima have to be computed, in form of fData object.

...

additional parameters.

which

logical flag specifying whether the grid values where maxima are fulfilled have to be returned too.

Value

If which = FALSE, the function returns a vector containing the maxima for each element of the functional dataset; if which = TRUE, the function returns a data.frame whose field value contains the values of maxima, and grid contains the grid points where maxima are reached.

See also

Examples

P = 1e3

grid = seq( 0, 1, length.out = P )

Data = matrix( c( 1 * grid,
                  2 *  grid,
                  3 * ( 0.5 - abs( grid - 0.5 ) ) ),
               nrow = 3, ncol = P, byrow = TRUE )

fD = fData( grid, Data )

maxima( fD, which = TRUE )
#>        grid    value
#> 1 1.0000000 1.000000
#> 2 1.0000000 2.000000
#> 3 0.4994995 1.498498