Operator sub-.fData to subset fData obejcts

This method provides an easy and natural way to subset a functional dataset stored in a fData object, without having to deal with the inner representation of fData class.

# S3 method for fData
[(fD, i, j, as_fData = TRUE)

Arguments

fD
the univariate functional dataset in form of fData object.

i
a valid expression to subset rows ( observations ) of the univariate functional dataset.

j
a valid expression to subset columns ( measurements over the grid ) of the univariate functional dataset (must be contiguous).

as_fData
logical flag to specify whether the output should be returned as an fData object containing the required subset or as a matrix of values, default is TRUE.

Value

The method returns either an fData object ( if as_fData = TRUE ) or a matrix ( if as_fData = FALSE ) containing the required subset ( both in terms of observations and measurement points ) of the univariate functional dataset.

See also

Examples

N = 20
P = 1e2

# One dimensional grid
grid = seq( 0, 1, length.out = P )

# Generating an exponential covariance function (see related help for more
# information )
C = exp_cov_function( grid, alpha = 0.3, beta = 0.4 )

# Generating a synthetic dataset with a gaussian distribution and
# required mean and covariance function:
fD = fData( grid,
            generate_gauss_fdata( N,
                                  centerline = sin( 2 * pi * grid ),
                                  Cov = C ) )

dev.new()

#> dev.new(): using pdf(file="Rplots288.pdf")

par( mfrow = c( 2, 2 ) )

# Original data
plot( fD )

# Subsetting observations
plot( fD[ c(1,2,3), , as_fData = TRUE ] )

# Subsetting measurements
plot( fD[ , 1 : 30 ] )

# Subsetting both observations and measurements
plot( fD[ 1 : 10, 50 : P ] )

# Subsetting both observations and measurements but returning a matrix
fD[ 1 : 10, 50 : P, as_fData = FALSE ]

#>              [,1]        [,2]       [,3]        [,4]         [,5]        [,6]
#>  [1,]  0.45163975  0.37063823  0.3100477  0.25947100  0.271919508  0.20443659
#>  [2,]  0.54942378  0.41994339  0.2835319  0.26769686  0.160754007  0.07296820
#>  [3,]  0.39471768  0.33572992  0.2721208  0.15278804  0.114013670  0.08602316
#>  [4,] -0.08730823 -0.08725450 -0.2260953 -0.28912206 -0.303327481 -0.35087289
#>  [5,]  0.19472265  0.19053426  0.1704277  0.07057526 -0.002006752 -0.08606790
#>  [6,] -0.23158056 -0.25510396 -0.2901949 -0.40726582 -0.533955701 -0.58462053
#>  [7,] -0.45554782 -0.54642656 -0.5678319 -0.61445229 -0.756612388 -0.80905215
#>  [8,]  0.01980222 -0.05244955 -0.1866059 -0.37196141 -0.444962366 -0.44921070
#>  [9,]  0.48494903  0.42917049  0.3232292  0.25073629  0.237750356  0.15629277
#> [10,] -0.40529387 -0.50984847 -0.6049663 -0.65352782 -0.697193512 -0.69171088
#>              [,7]         [,8]        [,9]       [,10]       [,11]       [,12]
#>  [1,]  0.15917045  0.063633468  0.10570446  0.11164466  0.08458548 -0.03116052
#>  [2,]  0.02847405 -0.060121497 -0.16237928 -0.19186947 -0.25413288 -0.27016887
#>  [3,]  0.05905041 -0.020839571 -0.02575229 -0.07173926 -0.03213704 -0.14442105
#>  [4,] -0.47944360 -0.577344516 -0.64822534 -0.75639970 -0.87428610 -0.96672252
#>  [5,] -0.13935315 -0.192102837 -0.27779018 -0.31264302 -0.36248924 -0.46691302
#>  [6,] -0.64473275 -0.731073197 -0.80742420 -0.74443906 -0.80189635 -0.90929695
#>  [7,] -0.93609509 -0.964073158 -1.02281946 -1.05503851 -1.12804997 -1.21755635
#>  [8,] -0.47683713 -0.610327200 -0.71791092 -0.75698013 -0.85575465 -0.84641648
#>  [9,]  0.02575339  0.008578354 -0.15199794 -0.20300581 -0.22764301 -0.27138489
#> [10,] -0.70146460 -0.738151676 -0.77968569 -0.89446263 -0.89216080 -0.96705414
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] -0.1355608 -0.2253586 -0.2464977 -0.2798601 -0.3200000 -0.4588439
#>  [2,] -0.1987710 -0.3475605 -0.3275519 -0.3625322 -0.3418008 -0.3949188
#>  [3,] -0.1367988 -0.2620394 -0.3549106 -0.4042829 -0.3838892 -0.3934441
#>  [4,] -1.0800137 -1.0138640 -1.0816137 -1.0905755 -1.1853816 -1.1629899
#>  [5,] -0.4020490 -0.5147375 -0.5578186 -0.6489731 -0.6706778 -0.6867790
#>  [6,] -0.9453166 -0.9342796 -0.9836248 -1.1072152 -1.0899089 -1.1213611
#>  [7,] -1.3162520 -1.3908347 -1.4996865 -1.5500154 -1.4728609 -1.4773727
#>  [8,] -0.8185812 -0.8590706 -0.9193546 -0.8860486 -0.9302936 -0.8918451
#>  [9,] -0.3015990 -0.2941591 -0.2980237 -0.4016144 -0.4485832 -0.4783176
#> [10,] -0.9443600 -0.9733464 -0.9638523 -1.0057680 -1.0014521 -0.8856540
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] -0.4464868 -0.4230790 -0.4387587 -0.4094491 -0.4687097 -0.3911988
#>  [2,] -0.3945027 -0.4053818 -0.4435408 -0.4259344 -0.4351814 -0.4513537
#>  [3,] -0.4871672 -0.4332343 -0.4841281 -0.4829880 -0.5052579 -0.5515869
#>  [4,] -1.3035126 -1.3229900 -1.3471009 -1.3450612 -1.3327199 -1.3011992
#>  [5,] -0.6984284 -0.6901598 -0.6564163 -0.6745324 -0.6613603 -0.7042289
#>  [6,] -0.9957848 -0.9239049 -0.9270030 -0.9330012 -0.9749046 -0.9974435
#>  [7,] -1.4894559 -1.5217985 -1.6401370 -1.6400128 -1.6211791 -1.6315351
#>  [8,] -0.8782873 -0.8661674 -0.8602821 -0.8687808 -0.8374690 -0.9006974
#>  [9,] -0.5123819 -0.4312863 -0.4113533 -0.4746321 -0.4921981 -0.5939221
#> [10,] -0.9037473 -0.9677276 -1.0954342 -1.0959420 -1.0887862 -1.1031351
#>            [,25]      [,26]      [,27]      [,28]      [,29]      [,30]
#>  [1,] -0.4204971 -0.3942928 -0.4103025 -0.3319525 -0.2732417 -0.3091692
#>  [2,] -0.4740043 -0.5016702 -0.4797935 -0.3441865 -0.4538745 -0.4710212
#>  [3,] -0.5773983 -0.5931789 -0.5856611 -0.5673815 -0.5314063 -0.5078678
#>  [4,] -1.2672877 -1.2759715 -1.2938832 -1.2460803 -1.1929202 -1.1932405
#>  [5,] -0.7047906 -0.8074951 -0.8000909 -0.7744974 -0.7290319 -0.7466713
#>  [6,] -1.0613729 -1.0915858 -1.0788055 -1.1387287 -1.1216466 -1.1529743
#>  [7,] -1.6349711 -1.5513204 -1.5730714 -1.5981106 -1.5795212 -1.5644449
#>  [8,] -0.9916457 -0.9884788 -0.9262910 -0.8893855 -0.8559249 -0.8419754
#>  [9,] -0.6012409 -0.6569219 -0.5781435 -0.5391701 -0.5422345 -0.5380943
#> [10,] -1.0520593 -0.9763232 -0.9683468 -0.8622206 -0.8475741 -0.6770441
#>            [,31]      [,32]      [,33]      [,34]      [,35]      [,36]
#>  [1,] -0.3316460 -0.2915805 -0.2131731 -0.1865330 -0.1771780 -0.2021677
#>  [2,] -0.5207832 -0.4863350 -0.5892088 -0.5558794 -0.4748813 -0.5059006
#>  [3,] -0.3643332 -0.3665296 -0.2757104 -0.3671645 -0.3005262 -0.2608449
#>  [4,] -1.1553798 -1.1925854 -1.2414472 -1.2890114 -1.2536518 -1.1799693
#>  [5,] -0.6871144 -0.6603323 -0.6051872 -0.5465181 -0.4787493 -0.5033484
#>  [6,] -1.2936864 -1.2706655 -1.2821107 -1.2265415 -1.2132955 -1.2673210
#>  [7,] -1.5533343 -1.5687233 -1.5036629 -1.4652018 -1.3640078 -1.2967047
#>  [8,] -0.7902039 -0.8705384 -0.7971657 -0.7271093 -0.7375903 -0.6714715
#>  [9,] -0.5805244 -0.6320162 -0.5602636 -0.4673300 -0.4045634 -0.3293259
#> [10,] -0.6833255 -0.5652141 -0.5324134 -0.5201687 -0.5255175 -0.5278270
#>             [,37]       [,38]       [,39]        [,40]       [,41]       [,42]
#>  [1,] -0.05286232 -0.01816841 -0.04208607  0.004537329  0.04580807  0.13164613
#>  [2,] -0.45155454 -0.33599414 -0.45065868 -0.439009748 -0.36090287 -0.32765595
#>  [3,] -0.21352540 -0.13801886 -0.11868965 -0.075698898 -0.02478917  0.06698437
#>  [4,] -1.14688953 -1.18601370 -1.09158065 -1.025594329 -1.01021831 -0.95723215
#>  [5,] -0.44470375 -0.40017118 -0.33252179 -0.399277009 -0.35734328 -0.30233028
#>  [6,] -1.25354351 -1.30761637 -1.19320021 -1.100189619 -1.01974683 -0.98560186
#>  [7,] -1.17831636 -1.17786227 -1.10318009 -1.029725400 -1.06723688 -0.99726791
#>  [8,] -0.69371536 -0.66254994 -0.66492189 -0.685176591 -0.58487364 -0.52576988
#>  [9,] -0.27446964 -0.25522256 -0.32637904 -0.239115330 -0.26463599 -0.20590980
#> [10,] -0.53936409 -0.44830885 -0.41182817 -0.338282482 -0.27195999 -0.24140376
#>            [,43]       [,44]       [,45]        [,46]       [,47]       [,48]
#>  [1,]  0.2478677  0.34412156  0.38379360  0.344409846  0.44135773  0.49290873
#>  [2,] -0.1651409 -0.12339146 -0.07979338 -0.008564613  0.04270243  0.13638185
#>  [3,]  0.1563871  0.14771608  0.23112082  0.360046728  0.44390373  0.52244257
#>  [4,] -0.8986581 -0.85350530 -0.77944027 -0.780650608 -0.65259464 -0.60312952
#>  [5,] -0.3151432 -0.28290788 -0.29517812 -0.203737729 -0.12899963 -0.09489081
#>  [6,] -0.9665240 -0.93744744 -0.86786440 -0.864628446 -0.81770195 -0.74916937
#>  [7,] -0.9807772 -0.98999541 -0.85630046 -0.835321826 -0.78197687 -0.73346242
#>  [8,] -0.3962469 -0.30265847 -0.19181080 -0.102736090 -0.11847714 -0.01381756
#>  [9,] -0.1501505 -0.09301395 -0.04060609 -0.015564462  0.05543072  0.11268757
#> [10,] -0.1982235 -0.10123488 -0.03679165  0.015868719  0.05062481  0.13331946
#>             [,49]       [,50]      [,51]
#>  [1,]  0.51981840  0.55652875  0.5996894
#>  [2,]  0.17935921  0.22195470  0.2068661
#>  [3,]  0.56413104  0.55401780  0.5613836
#>  [4,] -0.55399909 -0.54316188 -0.5183456
#>  [5,] -0.02577738  0.06553220  0.1379496
#>  [6,] -0.60107944 -0.55506339 -0.4129055
#>  [7,] -0.67770455 -0.68924916 -0.5918591
#>  [8,]  0.07101560  0.08741209  0.1732172
#>  [9,]  0.21801535  0.28933953  0.3007920
#> [10,]  0.16318451  0.26102026  0.3623919