Modified Epigraph Index for multivariate functional data

These functions compute the Modified Epigraph Index of elements of a multivariate functional dataset.

multiMEI(Data, weights = "uniform")

# S3 method for mfData
multiMEI(Data, weights = "uniform")

# S3 method for default
multiMEI(Data, weights = "uniform")

Arguments

Data

specifies the the multivariate functional dataset. It is either an object of class mfData or a list of 2-dimensional matrices having as rows the elements of that component and as columns the measurements of the functional data over the grid.

weights

either a set of weights (of the same length of Data ) or the string "uniform" specifying that a set of uniform weights (of value \(1 / L\), where \(L\) is the number of dimensions of the functional dataset and thus the length of Data) is to be used.

Value

The function returns a vector containing the values of MEI of each element of the multivariate functional dataset.

Details

Given a multivariate functional dataset composed of \(N\) elements with \(L\) components each, \(\mathbf{X_1} =( X^1_1(t),\) \(X^2_1(t), \ldots, X^L_1(t))\), and a set of \(L\) non-negative weights,

$$ w_1, w_2, \ldots, w_L, \qquad \sum_{i=1}^L w_i = 1,$$

these functions compute the MEI of each element of the functional dataset, namely:

$$ MEI( \mathbf{X_j} ) = \sum_{i=1}^{L} w_i MEI( X^i_j ), \quad \forall j = 1, \ldots N.$$

See also

Examples

N = 20
P = 1e3

grid = seq( 0, 10, length.out = P )

# Generating an exponential covariance function to be used to simulate gaussian
# functional data
Cov = exp_cov_function( grid, alpha = 0.2, beta = 0.8 )

# First component of the multivariate guassian functional dataset
Data_1 = generate_gauss_fdata( N, centerline = rep( 0, P ), Cov = Cov )

# First component of the multivariate guassian functional dataset
Data_2 = generate_gauss_fdata( N, centerline = rep( 0, P ), Cov = Cov )

mfD = mfData( grid, list( Data_1, Data_2 ) )

# Uniform weights
multiMEI( mfD, weights = 'uniform' )
#>  [1] 0.392375 0.460150 0.557175 0.605350 0.440975 0.467275 0.469650 0.484725
#>  [9] 0.533675 0.510700 0.566550 0.644900 0.586125 0.636600 0.587100 0.592850
#> [17] 0.401675 0.610700 0.468975 0.482475

# Non-uniform, custom weights
multiMEI( mfD, weights = c(2/3, 1/3) )
#>  [1] 0.4254167 0.5021667 0.5586333 0.6119333 0.4388333 0.4186167 0.4686333
#>  [8] 0.4743667 0.5116667 0.4520333 0.5224333 0.7040167 0.6215833 0.6776500
#> [15] 0.5466500 0.5892833 0.3786167 0.6069500 0.4426833 0.5478333