Operator sub-.fData to subset fData objects

This method provides an easy and natural way to subset a functional dataset stored in a fData object, without having to deal with the inner representation of fData class.

# S3 method for fData
[(fD, i, j, as_fData = TRUE)

Arguments

fD
the univariate functional dataset in form of fData object.

i
a valid expression to subset rows ( observations ) of the univariate functional dataset.

j
a valid expression to subset columns ( measurements over the grid ) of the univariate functional dataset (must be contiguous).

as_fData
logical flag to specify whether the output should be returned as an fData object containing the required subset or as a matrix of values, default is TRUE.

Value

The method returns either an fData object ( if as_fData = TRUE ) or a matrix ( if as_fData = FALSE ) containing the required subset ( both in terms of observations and measurement points ) of the univariate functional dataset.

See also

Examples

N = 20
P = 1e2

# One dimensional grid
grid = seq( 0, 1, length.out = P )

# Generating an exponential covariance function (see related help for more
# information )
C = exp_cov_function( grid, alpha = 0.3, beta = 0.4 )

# Generating a synthetic dataset with a gaussian distribution and
# required mean and covariance function:
fD = fData( grid,
            generate_gauss_fdata( N,
                                  centerline = sin( 2 * pi * grid ),
                                  Cov = C ) )

dev.new()
oldpar <- par(mfrow = c(1, 1))
par(mfrow = c(2, 2))

# Original data
plot(fD)

# Subsetting observations
plot(fD[c(1, 2, 3), , as_fData = TRUE])

# Subsetting measurements
plot(fD[, 1:30])

# Subsetting both observations and measurements
plot(fD[1:10, 50:P])

par(oldpar)

# Subsetting both observations and measurements but returning a matrix
fD[1:10, 50:P, as_fData = FALSE]

#>                [,1]        [,2]        [,3]         [,4]        [,5]
#>  [1,]  0.4117550337  0.32236708  0.30371229  0.260151400  0.26470166
#>  [2,] -0.9715184858 -1.00419190 -1.06512266 -1.043306764 -1.02989996
#>  [3,]  0.3473736190  0.31117670  0.24168356  0.155269472  0.07230493
#>  [4,] -0.0529144867 -0.06260534 -0.18372273 -0.201489062 -0.34098282
#>  [5,] -0.1780797657 -0.23921413 -0.25068789 -0.388319342 -0.48939567
#>  [6,] -0.6821966353 -0.70399376 -0.74663727 -0.875810642 -0.96385696
#>  [7,] -0.1221285864 -0.22489731 -0.24738156 -0.274683856 -0.30960530
#>  [8,]  0.3824777076  0.31260996  0.30240854  0.299287149  0.17530358
#>  [9,]  0.3405459867  0.30156149  0.27174739  0.100547766  0.01594609
#> [10,] -0.0006634077 -0.04220848 -0.07844234 -0.008395895 -0.20141872
#>              [,6]        [,7]         [,8]        [,9]      [,10]      [,11]
#>  [1,]  0.17778894  0.09059415 -0.043926114 -0.07322473 -0.1798496 -0.1162265
#>  [2,] -1.07051077 -1.14734769 -1.139256613 -1.17888073 -1.2091814 -1.2678685
#>  [3,] -0.02812209 -0.08913366 -0.113651162 -0.22378609 -0.4232280 -0.5395075
#>  [4,] -0.44858911 -0.43073633 -0.459967436 -0.54307438 -0.5688011 -0.6079854
#>  [5,] -0.53332211 -0.56688782 -0.614764081 -0.68372013 -0.7056252 -0.7045901
#>  [6,] -0.99748877 -1.07619490 -1.104783414 -1.15397751 -1.1770430 -1.2268812
#>  [7,] -0.30794188 -0.33817898 -0.390488740 -0.37717569 -0.4385441 -0.4470224
#>  [8,]  0.04977131  0.05232310  0.003729727 -0.06754408 -0.1386045 -0.1647982
#>  [9,] -0.06195210 -0.15700996 -0.156346602 -0.24803851 -0.2483973 -0.3465733
#> [10,] -0.30236413 -0.32872937 -0.344244165 -0.39591317 -0.3661461 -0.5358538
#>            [,12]      [,13]      [,14]      [,15]      [,16]      [,17]
#>  [1,] -0.1567453 -0.1811200 -0.3187751 -0.3918567 -0.5635107 -0.6097068
#>  [2,] -1.3216648 -1.3313727 -1.3982080 -1.4091936 -1.4394230 -1.4077045
#>  [3,] -0.6071868 -0.6601090 -0.7594773 -0.7519240 -0.8712149 -0.9305934
#>  [4,] -0.6687814 -0.6432351 -0.6210015 -0.6417276 -0.6756630 -0.6436326
#>  [5,] -0.8180155 -0.8998448 -0.9267785 -0.9263956 -0.9392964 -0.9174525
#>  [6,] -1.2914846 -1.3424930 -1.4562270 -1.4177479 -1.4261179 -1.4508744
#>  [7,] -0.5095515 -0.5057731 -0.5204100 -0.5874304 -0.5745256 -0.5337292
#>  [8,] -0.2191443 -0.2378587 -0.3129006 -0.3444803 -0.4346360 -0.4200229
#>  [9,] -0.4241594 -0.4764108 -0.5039370 -0.5486235 -0.5925565 -0.7238532
#> [10,] -0.6126065 -0.6527413 -0.7058915 -0.8134147 -0.8643366 -0.8410498
#>            [,18]      [,19]      [,20]      [,21]      [,22]      [,23]
#>  [1,] -0.6119615 -0.5858906 -0.5668783 -0.5889519 -0.7437500 -0.7488450
#>  [2,] -1.4566849 -1.4558040 -1.4548010 -1.3753773 -1.3457303 -1.2950167
#>  [3,] -1.0432287 -1.0480186 -1.0161905 -1.0316634 -0.9772514 -1.0245949
#>  [4,] -0.6869172 -0.8053629 -0.9249606 -1.0150202 -0.9708729 -1.0519170
#>  [5,] -0.8551686 -0.9437411 -0.9563884 -0.9748660 -0.9535126 -0.9106065
#>  [6,] -1.4561258 -1.5284677 -1.5409764 -1.5569443 -1.6617858 -1.6776058
#>  [7,] -0.5264472 -0.5477646 -0.6226420 -0.6207793 -0.7053485 -0.7864337
#>  [8,] -0.4589901 -0.5431607 -0.5482069 -0.6006526 -0.5801518 -0.6587052
#>  [9,] -0.7509913 -0.7865587 -0.8696694 -0.8990953 -0.9469473 -0.9761153
#> [10,] -0.8503600 -0.8487580 -0.9400607 -0.9727684 -1.0283788 -1.0472653
#>            [,24]      [,25]      [,26]      [,27]      [,28]      [,29]
#>  [1,] -0.8450961 -0.8678270 -0.8133462 -0.8379223 -0.8408420 -0.8100282
#>  [2,] -1.3810252 -1.3801420 -1.3312113 -1.2997768 -1.3221006 -1.3009615
#>  [3,] -0.9656401 -0.9327070 -1.0173822 -1.0384794 -1.0960465 -1.0295004
#>  [4,] -1.0644573 -1.0327893 -1.1115067 -1.1432135 -1.1466277 -1.1578301
#>  [5,] -0.9081034 -0.9643857 -1.0177573 -1.0510301 -1.1064452 -1.1350065
#>  [6,] -1.6574922 -1.6926019 -1.6848835 -1.5944646 -1.5953285 -1.5412278
#>  [7,] -0.8270131 -0.7373801 -0.7742520 -0.6741507 -0.5831570 -0.6172949
#>  [8,] -0.6381852 -0.5878621 -0.6151577 -0.5388613 -0.5650724 -0.5429943
#>  [9,] -0.9821314 -1.0705737 -1.0473901 -1.0058579 -1.0231792 -1.0520359
#> [10,] -1.0614271 -1.1296454 -1.1519051 -1.1359654 -1.1483919 -1.1540206
#>            [,30]      [,31]      [,32]      [,33]      [,34]      [,35]
#>  [1,] -0.8182153 -0.7702367 -0.7630581 -0.7308462 -0.7956429 -0.7931432
#>  [2,] -1.2717633 -1.3313007 -1.4060649 -1.2835952 -1.2713067 -1.2974311
#>  [3,] -1.0996887 -1.1105239 -1.1314006 -1.0403864 -1.0541750 -0.9059475
#>  [4,] -1.1046607 -1.1467267 -1.1409729 -0.9970027 -0.9577703 -0.9112958
#>  [5,] -1.1307136 -1.0515398 -1.0503914 -1.0540172 -0.9970704 -0.9632043
#>  [6,] -1.4715968 -1.3968560 -1.4314495 -1.3976326 -1.3697404 -1.3318868
#>  [7,] -0.6356417 -0.6115167 -0.6263126 -0.6602754 -0.6546210 -0.5984072
#>  [8,] -0.6366414 -0.6285086 -0.5301458 -0.5185070 -0.4665022 -0.4087009
#>  [9,] -1.0347310 -1.0734174 -0.9447674 -0.8896672 -0.8510114 -0.8428074
#> [10,] -1.1149260 -1.0468263 -1.0179796 -1.0259178 -1.0350501 -0.9620754
#>            [,36]      [,37]      [,38]      [,39]      [,40]      [,41]
#>  [1,] -0.7927431 -0.7696883 -0.7932381 -0.6810097 -0.7570988 -0.7589080
#>  [2,] -1.2983472 -1.2111303 -1.2169525 -1.1782330 -1.1043915 -1.0778317
#>  [3,] -0.8668407 -0.8030211 -0.7175184 -0.6902975 -0.6624334 -0.6966289
#>  [4,] -0.8753610 -0.8325727 -0.7878939 -0.7119321 -0.6433794 -0.5180420
#>  [5,] -0.8128825 -0.7432216 -0.6652401 -0.6546764 -0.6554118 -0.6090906
#>  [6,] -1.2859179 -1.2401577 -1.1975600 -1.1643115 -1.0393426 -1.0085208
#>  [7,] -0.5971221 -0.5837701 -0.4760343 -0.5219596 -0.4932354 -0.5143135
#>  [8,] -0.3467352 -0.3758258 -0.2685073 -0.2917724 -0.2561237 -0.1633503
#>  [9,] -0.8253739 -0.9098637 -0.8846608 -0.7583994 -0.7820653 -0.7779955
#> [10,] -0.8856830 -0.8441057 -0.8192540 -0.8139600 -0.7825331 -0.7019789
#>            [,42]        [,43]       [,44]       [,45]       [,46]       [,47]
#>  [1,] -0.6278189 -0.555819087 -0.49080269 -0.46774363 -0.44626604 -0.37620045
#>  [2,] -1.0558127 -1.014167790 -0.90119193 -0.72027014 -0.65461273 -0.48733715
#>  [3,] -0.6163693 -0.532359214 -0.48059962 -0.47940620 -0.46520203 -0.42813923
#>  [4,] -0.4684488 -0.366843077 -0.31144455 -0.28442610 -0.17572315 -0.08996673
#>  [5,] -0.5315996 -0.572821501 -0.43878195 -0.33091959 -0.25702342 -0.27468002
#>  [6,] -1.0193169 -1.005121017 -0.89310849 -0.86214193 -0.84860042 -0.78843216
#>  [7,] -0.4193449 -0.412117010 -0.35140116 -0.30435473 -0.23136802 -0.12569304
#>  [8,] -0.1374741 -0.005986014 -0.02146941  0.08079984  0.08755569  0.11242275
#>  [9,] -0.7078459 -0.695681788 -0.70646239 -0.59679385 -0.58044546 -0.51725778
#> [10,] -0.6010839 -0.532185735 -0.53679659 -0.43941184 -0.38045713 -0.29772540
#>             [,48]       [,49]       [,50]       [,51]
#>  [1,] -0.24162523 -0.22770116 -0.10213513 -0.06871722
#>  [2,] -0.30510374 -0.16847084 -0.09170299  0.03959049
#>  [3,] -0.38290420 -0.32215802 -0.21891142 -0.18286156
#>  [4,]  0.08779548  0.06005955  0.13393035  0.12315454
#>  [5,] -0.17193915 -0.07539826 -0.02527484  0.04170011
#>  [6,] -0.72502690 -0.63319681 -0.60645801 -0.47737938
#>  [7,] -0.01650311  0.09021269  0.25968557  0.38148633
#>  [8,]  0.12677444  0.18567636  0.26255110  0.31010246
#>  [9,] -0.56891294 -0.52324810 -0.43068016 -0.31874636
#> [10,] -0.22615644 -0.18186528 -0.12091505 -0.02520150